+++ src/selectors/controlSelectors.tsx
@@ -41,7 +41,7 @@ import { Layers } from '../components/ol/layer/Layers';
 import { findIndexCloseTo } from '../util/find';
 import { Server } from '../model/server';
 import { MapGroup, maps, MapSource } from '../util/maps';
-import { getTileAccess } from '../config';
+import { getBranding, getTileAccess } from '../config';
 
 export const selectedDatasetIdSelector = (state: AppState) => state.controlState.selectedDatasetId;
 export const selectedVariableNameSelector = (state: AppState) => state.controlState.selectedVariableName;
@@ -386,19 +386,23 @@ export const selectedDatasetRgbLayerSelector = createSelector(
     }
 );
 
+const DEFAULT_FILL_OPACITY = getBranding().polygonFillOpacity || 0.25;
 
 const DEFAULT_STROKE = new OlStrokeStyle({
                                              color: [200, 0, 0, 0.75],
                                              width: 1.25
                                          });
+
 const DEFAULT_FILL = new OlFillStyle({
-                                         color: [255, 0, 0, 0.25]
+                                         color: [255, 0, 0, DEFAULT_FILL_OPACITY]
                                      });
+
 const DEFAULT_IMAGE = new OlCircle({
                                        fill: DEFAULT_FILL,
                                        stroke: DEFAULT_STROKE,
                                        radius: 6
-                                   })
+                                   });
+
 const DEFAULT_PLACE_GROUP_STYLE = new OlStyle(
     {
         image: DEFAULT_IMAGE,